Add files via upload
I tried making the Mars Terrain in blender and I was able to. I converted the mars terrain into an STL because I wanted to use this tool called numpy-stl to convert it into a npy file and use it in the code but I have been having some problems with it as now the setup file is giving me errors on the dimensions of the mars terrain. he dimensions I made it were 64 x 64 x 1.83 meters.